requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
typer==0.12.5
jinja2==3.1.4
pyyaml==6.0.2
markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
weasyprint==61.2
beautifulsoup4==4.12.3
lxml==5.3.0
matplotlib==3.9.2
pydantic==2.9.2
python-multipart==0.0.9
language-tool-python==2.8.1
requests==2.32.3
pytest==8.3.3
pytest-cov==5.0.0
mdit-py-plugins==0.4.0
click==8.1.7
pydyf==0.9.0